Organization/tidying
Organize app into containers, components and utils, as the src
directory was becoming a bit overloaded.

Ensure tests of a single function/component are wrapped in a describe
block which references the function/component.
