+++ src/App.js
@@ -2,15 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Button from 'material-ui/Button';
-import { map } from 'ramda';
 import { initializeSearch } from './actions';
-import formatDepartureData from './formatDepartureData';
-import AppBox from './AppBox';
-import AppHead from './AppHead';
-import Departure from './Departure';
+import AppBox from './components/AppBox';
+import AppHead from './components/AppHead';
+import DeparturesList from './containers/DeparturesList';
 import './App.css';
 
-const App = ({ dispatch, departures }) => {
+const App = ({ dispatch }) => {
   const searchNow = () => {
     return dispatch(initializeSearch());
   };
@@ -22,25 +20,13 @@ const App = ({ dispatch, departures }) => {
       <Button className="button" onClick={searchNow}>
         Search Now
       </Button>
-      <div className="App-body">
-        {map(departureInfo => <Departure {...departureInfo} />, departures)}
-      </div>
+      <DeparturesList />
     </AppBox>
   );
 };
 
-const mapStateToProps = ({ tripInformation }) => {
-  const departures = formatDepartureData(tripInformation);
-  return { departures };
-};
-
 App.propTypes = {
   dispatch: PropTypes.func.isRequired,
-  departures: PropTypes.arrayOf(PropTypes.object),
 };
 
-App.defaultProps = {
-  departures: [],
-};
-
-export default connect(mapStateToProps)(App);
+export default connect()(App);
