+++ src/components/AppHead.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from './images/busbudLogo.png';
+import logo from '../images/busbudLogo.png';
 
 const AppHead = () => {
   return (
@@ -9,7 +9,7 @@ const AppHead = () => {
         <h1 className="App-title">Get to Osheaga!</h1>
       </header>
       <p className="App-subheader">
-      Find tickets from New York to Montréal on August 2nd, 2018:
+        Find tickets from New York to Montréal on August 2nd, 2018:
       </p>
     </div>
   );
